backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data validation
pydantic==2.7.4
pydantic-settings==2.4.0

# Document processing - Main parser
docling>=2.67.0,<3.0.0
docling-core>=2.50.1,<3.0.0

# Document processing - Fallback loaders
PyMuPDF==1.23.8
PyPDF2==3.0.1
unstructured==0.11.2
python-docx>=1.1.2,<2.0.0
# textract removed - conflicts with beautifulsoup4>=4.12.3 (requires ~=4.8.0)
# DOC files use antiword or python-docx as fallback
openpyxl>=3.1.5,<4.0.0
python-pptx>=1.0.2,<2.0.0
beautifulsoup4>=4.12.3,<5.0.0
lxml>=4.0.0,<7.0.0
pandas>=2.1.4,<3.0.0
ebooklib==0.18
extract-msg==0.48.4
webvtt-py==0.4.6
langchain==0.3.15
langchain-community==0.3.14
langchain-text-splitters==0.3.4
langchain-core==0.3.31
langchain-openai==0.2.14

# AI and Embeddings
openai==1.109.1
boto3==1.34.3
sentence-transformers==2.2.2
huggingface-hub>=0.23,<1
tiktoken==0.12.0

# Vector stores
pymilvus==2.3.4
pinecone-client==2.2.4
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests>=2.32.2,<3.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
reportlab==4.0.7
